[BugFix] Optimise moe lora allgather indieces  by change argsort input tensor type (#16486)
### What this PR does / why we need it?

This PR optimizes MoE LoRA routing recovery in the AllGather path by
casting `torch.abs(expanded_row_idx)` to `torch.float32` before calling
`torch.argsort`.

On Ascend NPUs, `torch.argsort` runs faster on `float32` tensors than on
`torch.long` (`int64`) tensors. This change reduces the sorting overhead
when recovering the inverse permutation used to align expert IDs and
LoRA slots with dispatched tokens. The permutation returned by `argsort`
remains an integer index tensor.

### Does this PR introduce _any_ user-facing change?

This is a performance optimization for MoE LoRA inference using
AllGather dispatch. No API or configuration changes are introduced.

### How was this patch tested?

TODO: Add routing correctness test results and NPU benchmark
measurements, including the float32 conversion overhead.


- vLLM main:
https://github.com/vllm-project/vllm/commit/a97dacb7106ee49f39f3d1fc6ae1800ff724e01d
```text
T       32      128     512     2048    4096
Long    ~0.073  ~0.139  ~0.421  ~1.77   ~3.64   (ms)
float32 ~0.060  ~0.065  ~0.074  ~0.117  ~0.174  (ms)
```

+++ vllm_ascend/lora/fused_moe.py
@@ -199,7 +199,7 @@ def _recover_moe_lora_routing_allgather(lora_context, expanded_row_idx, topk_ids
     ``.item()``/data-dependent host sync.
     """
     top_k = lora_context.top_k
-    expanded = torch.abs(expanded_row_idx)
+    expanded = torch.abs(expanded_row_idx).to(torch.float32)
     inv_perm = torch.argsort(expanded)
     expert_per_row = topk_ids.reshape(-1)[inv_perm].to(torch.long)
